requirements/local.txt
angquick==1.0
dj-database-url==0.4.1
Django==1.8.14
djangorestframework==3.4.7
gunicorn==19.6.0
Unipath==1.1
whitenoise==3.2.2
